Move version_utils to _internal | chore
Move version_utils to `_internal` so that it can be used my onnxscript

[ghstack-poisoned]
