+++ onnxscript/tests/functions/onnxfns1A_test.py
@@ -3,7 +3,8 @@ import unittest
 import onnx
 import pytest
 
-from onnxscript.tests.common import onnx_script_test_case, version_utils
+from onnxscript._internal import version_utils
+from onnxscript.tests.common import onnx_script_test_case
 from onnxscript.tests.models import onnxfns1A
